Update of Memento Pattern
Changes -

1. readMe
Updated the readMe with an example of selecting a original state,
saving it, modifying it then restoring. It functions like a
refresh/apply, where we are allowed to restore to a previous state even
if we have made changes.

The example used is deciding on the colour of a shirt, we save the
colour when required and restore it if we have changed our minds at a
later stage.

2. Project Code

Implemented the t-shirt example, where the originator is the t-shirt
object that we want to change the state of. It creates a memento of the
saved colour (state) that we restore when required.