Delivery of State Pattern
Changes -

1. readMe
Provided readMe to define the state pattern, its uses and a description
of the code example.

2. Code Example
Provided code example is of a traffic light system. In this demo, the
lights represent the states and by changing the context state allows us
to change the colour of the light displayed.